requirements.txt
sentence-transformers
faiss-cpu
transformers
torch
datasets
numpy
pandas
matplotlib
seaborn
tqdm
scikit-learn